==> notebooks/1.3 Build Index.ipynb <==
# Setup Environment

from working_dir import set_working_dir

set_working_dir()

%load_ext autoreload
%autoreload 2

# Import Libraries

import metapy

# Build Index

idx = metapy.index.make_inverted_index('config.toml')

print(f"# Sentence: {idx.num_docs()},  Avg length: {idx.avg_doc_length()}")

query = metapy.index.Document() # Define query
query.content('washed') # search word "washed"

# Check the ranking method in the object
object_methods = [method_name for method_name in dir(metapy.index)
                  if callable(getattr(metapy.index, method_name))]

object_methods

# Define ranker
ranker = metapy.index.AbsoluteDiscount()

# Show top 2 docs containing word 'washed'
top_docs = ranker.score(idx, query, num_results=2)
top_docs

# Show full sentence
for num, (d_id, _) in enumerate(top_docs):
    content = idx.metadata(d_id).get('content')
    print("{}. {}...\n".format(num + 1, content))
