==> notebooks/2.2 Interesting Words.ipynb <==
# Setup Environment

from working_dir import set_working_dir

set_working_dir()

%load_ext autoreload
%autoreload 2

# Import libraries

import pandas as pd
from sentence import Sentences, Sentence

# Word Frequency
From word frequency file, we have 5050 ranked words.

file_path = "data/wordFrequency.xlsx"

wf_df = pd.read_excel(io=open(file_path, 'rb'), sheet_name='4 forms (219k)', engine="openpyxl")

len(wf_df)

wf_df.head()

wf_df.columns

# Sentences
We load sentences from the cleaned data.

sentences_df = pd.read_csv(f"data/cbtest_NE_train_cleaned.csv")
sentences_df

sentences = Sentences([Sentence(s) for s in sentences_df["sentence"].tolist()])

len(sentences)

# Pick interesting words
Based on `1.2 Word to Sentence.ipynb` exploration, we would like to pick the words that have at least 10 example sentences but less than 1000 example sentences. The words that have greater than 1000 example sentences are common words, which are not so interesting.

n_examples_lower_bound = 10
n_examples_upper_bound = 1000

def count_containing_sentence(word):
    sentences_containing_word = sentences.filter_by_word(word)
    return len(sentences_containing_word)

word_sentence_df = pd.DataFrame(columns=["word", "rank", "sentence_count"])
word_sentence_df["word"] = wf_df["word"]
word_sentence_df["rank"] = wf_df["rank"]
word_sentence_df["sentence_count"] = word_sentence_df["word"].apply(count_containing_sentence)

len(word_sentence_df)

len(word_sentence_df[word_sentence_df["sentence_count"] >= n_examples_lower_bound])

interesting_word_sentence_df = word_sentence_df[(word_sentence_df["sentence_count"] >= n_examples_lower_bound) & (word_sentence_df["sentence_count"] <= n_examples_upper_bound)]

len(interesting_word_sentence_df)

interesting_word_sentence_df.head()

# Write interesting words
We exacted 3049 interesting words out of 5050 words.

interesting_word_sentence_df.to_csv('data/interesting_words.csv', index=False)
