==> notebooks/2.3 Word Samples.ipynb <==
# Setup Environment

from working_dir import set_working_dir

set_working_dir()

%load_ext autoreload
%autoreload 2

# Import libraries

import pandas as pd
import numpy as np

# Pick word samples
From interesting word list, we try to get some words from each difficulty levels as sample words.

df = pd.read_csv("data/interesting_words.csv")

df.head()

rank_below1k = df["rank"] <= 1000
rank_1k_2k = (df["rank"] > 1000) & (df["rank"] <= 2000)
rank_2k_3k = (df["rank"] > 2000) & (df["rank"] <= 3000)
rank_3k_4k = (df["rank"] > 3000) & (df["rank"] <= 4000)
rank_above4k = df["rank"] > 4000

pd.concat([df[rank_below1k].head(1), df[rank_below1k].tail(1)])

pd.concat([df[rank_1k_2k].head(1), df[rank_1k_2k].tail(1)])

pd.concat([df[rank_2k_3k].head(1), df[rank_2k_3k].tail(1)])

pd.concat([df[rank_3k_4k].head(1), df[rank_3k_4k].tail(1)])

pd.concat([df[rank_above4k].head(1), df[rank_above4k].tail(1)])

# Sampling demo words from each rank bands
We randomly pick 5 words from different rank bands as example words.

np.random.seed(42)

def sample(dfs: [], n:int):
    sample_dfs = []
    for df in dfs:
        sample_dfs.append(df.sample(n=n))
    return pd.concat(sample_dfs)

sample_df = sample(
    dfs=[df[rank_below1k], df[rank_1k_2k], df[rank_2k_3k], df[rank_3k_4k], df[rank_above4k]],
    n=1)

sample_df

# Write demo words to file

sample_df_final = sample_df

sample_df_final.to_csv('data/sample_words.csv', index=False) 

# Sampling evaluation words from each rank bands
We randomly pick 10 words from different rank bands as evaluation words.

np.random.seed(52)

eval_df = sample(
    dfs=[df[rank_below1k], df[rank_1k_2k], df[rank_2k_3k], df[rank_3k_4k], df[rank_above4k]],
    n=1)

eval_df

# Write eval words to file

eval_df_final = eval_df

eval_df_final.to_csv('data/eval_words.csv', index=False)
