==> notebooks/3.3 Combined Difficulty.ipynb <==
# Setup Environment

from working_dir import set_working_dir

set_working_dir()

%load_ext autoreload
%autoreload 2

# Import Libraries

import pandas as pd
from sentence_ranker import SentenceRanker, Inclusion
from pos_difficulty_ranker import POSDifficultyRanker
from word_frequency import WordFrequencies
from difficulty_ranker import DifficultyRankers
from word_difficulty_ranker import WordDifficultyRanker

%matplotlib inline

pd.set_option('max_colwidth', 400)

# Combined Difficulty
We experiment and test out combined difficulty(combine word difficulty and POS difficulty with equal weight 0.5) with some example sentences.

word_frequencies = WordFrequencies.read_excel(file_path="data/wordFrequency.xlsx")

word_difficulty_ranker = WordDifficultyRanker(word_frequencies=word_frequencies)
pos_difficulty_ranker = POSDifficultyRanker()
difficulty_rankers = DifficultyRankers(
        rankers=[word_difficulty_ranker, pos_difficulty_ranker],
        weights=[0.5, 0.5])
ranker = SentenceRanker(difficulty_rankers)

examples = [
    "What do you really think about it?",
    "I don’t care about your past.",
    "What do you think about that issue?",
    "He genuinely cares about me.",
    "Do you have any information about the project?",
    "Our house is worth about 150,000 dollars.",
    "Don’t worry too much about unimportant things.",
    "John's mother abandoned him in cold heart after he was born",
]

ranked_sentences = ranker.rank(examples)

ranked_sentences.sort_by_difficulty()

df = ranked_sentences.to_pandas(inclusion = Inclusion.All)

df
