==> notebooks/4.1 Evaluating Sample Words.ipynb <==
# Setup Environment

from working_dir import set_working_dir

set_working_dir()

%load_ext autoreload
%autoreload 2

# Import libraries

import numpy as np
from word_app import create_app
import pandas as pd
from sentence_ranker import Inclusion

pd.set_option('max_colwidth', 400)

# Sample words

sample_words_df = pd.read_csv("data/sample_words.csv")

sample_words_df.head()

sample_words = sample_words_df["word"].tolist()

sample_words

# Evaluating sample words

app = create_app()

def evaluate_word(word: str):
    ranked_sentences = app.search_for_sentences(word, n_results=100)
    ranked_sentences.sort_by_difficulty()
    return ranked_sentences.to_pandas(inclusion = Inclusion.All)

print(sample_words[0])
evaluate_word(sample_words[0])

print(sample_words[1])
evaluate_word(sample_words[1])

print(sample_words[2])
evaluate_word(sample_words[2])

print(sample_words[3])
evaluate_word(sample_words[3])

print(sample_words[4])
evaluate_word(sample_words[4])
