==> notebooks/4.1 Evaluation.ipynb <==
# Setup Environment

from working_dir import set_working_dir

set_working_dir()

%load_ext autoreload
%autoreload 2

# Import libraries

from word_app import create_app
import pandas as pd
from ndcg import NDCG

pd.set_option('max_colwidth', 400)

# Evaluation query list

qeuries_df = pd.read_csv("data/eval_queries.csv")

qeuries_df

# Cranfield Evaluation(NDCG)
We do evaluation based on the sample queries we have prepared above.

app = create_app()

ndcg = NDCG("data/eval_queries.csv")

words = qeuries_df["word"].unique().tolist()
words

scores = {}
for word in words:
    word_result = app.search(word, n_results=5)
    sentences = [str(s) for s in word_result.ranked_sentences]
    scores[word] = round(ndcg.score(word, sentences), 3)
scores_df = pd.DataFrame(scores.items(), columns=['word', 'score'])

print(f"Average Score: {round(scores_df['score'].mean(), 3)}")

scores_df
