==> notebooks/5.1 User.ipynb <==
# Setup Environment

from working_dir import set_working_dir

set_working_dir()

%load_ext autoreload
%autoreload 2

# Import Libraries

from word_app import create_app
import pandas as pd
from sentence_ranker import Inclusion
from IPython.core.display import display, HTML

pd.set_option('max_colwidth', 400)


# Project Scope

Given the fact that many high quality data sources are not freely available online, the project only focus on the most frequent 5,000 english words found in "data/wordFrequency.xlsx"

# Demo

app = create_app()

def search_and_display_word(word: str, n_results=10):
    word_result = app.search(word, n_results)
    return display(HTML(word_result.to_html()))

def evaluate_word(word: str):
    ranked_sentences = app.search_for_sentences(word)
    ranked_sentences.sort_by_difficulty()
    return ranked_sentences.to_pandas(inclusion = Inclusion.All)

search_and_display_word(word="corner", n_results=10)

search_and_display_word(word="newspapers", n_results=11)

search_and_display_word(word="STORY", n_results=5)


evaluate_word("story")
